Update minio to explicitly specify console address
This commit adds changes to explicitly specify minio console address as in the latest version(s), minio is now exposing 2 ports for different reasons. For UI we have console address and for minio server port, we have a separate port. If minio console address is not explicitly specified, it results in minio randomly choosing a port which it wants to use for minio console which results in inconsistent behavior as port forwarding gets broken with the new dynamic port not being mapped